remove general device
