+++ CardScan/CardScan/Classes/Api/ScanStatistics.swift
@@ -8,20 +8,42 @@
 import Foundation
 
 struct ScanStatisticsPayload: Encodable {
-    let device = Device()
+    let device = ScanStatisticDevice()
     let scanStats: ScanStatistics
-    let sdkVersion = AppInfoUtils.getSdkVersion()
-    
+
     enum CodingKeys: String, CodingKey {
         case scanStats = "scan_stats"
-        case sdkVersion = "sdk_version"
     }
     
     func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(scanStats, forKey: .scanStats)
-        try container.encode(sdkVersion, forKey: .sdkVersion)
         try device.encode(to: encoder)
+    }
+}
+
+struct ScanStatisticDevice: Encodable {
+    let locale: String? = DeviceUtils.locale
+    let build: String = DeviceUtils.build
+    let osVersion: String = DeviceUtils.osVersion
+    let platform: String = DeviceUtils.platform
+    let sdkVersion = AppInfoUtils.getSdkVersion()
+    
+    enum CodingKeys: String, CodingKey {
+        case locale = "device_locale"
+        case build = "build"
+        case osVersion = "os"
+        case platform = "platform"
+        case sdkVersion = "sdk_version"
+    }
+    
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(locale, forKey: .locale)
+        try container.encode(build, forKey: .build)
+        try container.encode(osVersion, forKey: .osVersion)
+        try container.encode(platform, forKey: .platform)
+        try container.encode(sdkVersion, forKey: .sdkVersion)
     }
 }
